debug_info: Support LLVM 16
This patch adds `llvm16-0` to feature gates in the call to
`LLVMDIBuilderCreateCompileUnit` in
`DebugInfoBuilder::create_compile_unit()`. It also adds the flag to the
arguments `sysroot` and `sdk` in the function declaration and to all of
the calls to it, as well as to `DebugInfoBuilder::new()`.

Furthermore, it adds the following enum values to
`flags::DwarfSourceLanguage`:
- `Kotlin`
- `Zig`
- `Crystal`
- `CPlusPlus17`
- `CPlusPlus20`
- `C17`
- `Fortran18`
- `Ada2005`
- `Ada2012`